requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
jinja2
openai
python-dotenv
requests
psycopg2-binary # Mantenuto per flessibilità futura, ma non strettamente necessario per SQLite puro.
email-validator
